+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- 2026-04-28 | ✨ feat(access): implement HU-019 bylaws hybrid mode
 - 2026-04-19 | ✨ feat(media): integrate image picker in news and profile
 - 2026-04-18 | ✨ feat(access): implement admin users management
 - 2026-04-18 | ✨ feat(media): add shared image pipeline and naming